refactor: move `config` to `config/environment`
