+++ src/index.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import { Sync } from './sync.js';
-import { Config } from './config.js';
+import { Config } from './config/environment.js';
 
 do {
   await Sync.perform();
